Transaction logic

+++ FoodOrderSystem.Models/DTOs/Profile/GetUserResponseDto.cs
@@ -12,7 +12,7 @@
         public string? Address { get; set; }
         public string FullName { get; set; } = null!;
         public string? Avatar { get; set; }
-        public double WalletBalance { get; set; } = 0;
+        public decimal WalletBalance { get; set; } = 0;
         public string RoleName { get; set; } = "Student"; // "Student", "ShopOwner", or "Admin"
         public string? UserName { get; set; }
         public string? BirthDate { get; set; }
